requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.6
Python-EasyConfig==0.1.7
pytz==2022.1
PyYAML==6.0
Resource==0.2.1
six==1.16.0
SQLAlchemy==1.4.34
Werkzeug==2.1.1
